master

+++ resources/views/master/user/index.blade.php
@@ -1,4 +1,4 @@
-@include('_include.admin_header',
+@include('_include.master.header',
 	[
 		'id'				=> 'master_user',
 	]
@@ -16,6 +16,13 @@
 	@include('_include.alert_message', ["message" => (isset($message) ? $message : session("message")), "alert_type" => (isset($alert_type) ? $alert_type : session("alert_type"))])
 @endif
 
+<br>
+<div id="divAlertBox" class="alert w3-hide">
+	<span class="closebtn">&times;</span>
+	<div id="divMessage"></div>
+</div>
+<br>
+
 <div class="w3-row">
 	<button type="button" class="w3-button w3-brown" ng-click="reset()"><span class="glyphicon glyphicon-list"></span></button>&nbsp;
 	<a href="{{ $data['url_pattern'] }}/add" class="w3-button w3-brown"><span class="glyphicon glyphicon-plus"></span></a>
@@ -28,22 +35,22 @@
 	<input type="hidden" id="permission_flag" name="permission_flag" value="">
 	<input type="hidden" id="data_source_url" value="/api{{ $data['url_pattern'] }}">
 	@endif
+	<input type="hidden" id="data_source_url_delete" value="/api{{ $data['url_pattern'] }}">
 
 	<div class="w3-bar w3-light-gray">
-		<a class="w3-bar-item w3-button" href="{{ $data['url_pattern'] }}?permission=master">Master</a>
-		<a class="w3-bar-item w3-button" href="{{ $data['url_pattern'] }}?permission=other">Other users</a>
+		<a class="w3-bar-item w3-button {{ (!isset($data['permission_flag'])) ? 'w3-gray' : '' }}" href="{{ $data['url_pattern'] }}">All</a>
+		<a class="w3-bar-item w3-button {{ (isset($data['permission_flag']) && ($data['permission_flag'] == 'Master')) ? 'w3-gray' : '' }}" href="{{ $data['url_pattern'] }}?permission=Master">Master</a>
+		<a class="w3-bar-item w3-button {{ (isset($data['permission_flag']) && ($data['permission_flag'] == 'Owner')) ? 'w3-gray' : '' }}" href="{{ $data['url_pattern'] }}?permission=Owner">Owner</a>
+		<a class="w3-bar-item w3-button {{ (isset($data['permission_flag']) && ($data['permission_flag'] == 'Other')) ? 'w3-gray' : '' }}" href="{{ $data['url_pattern'] }}?permission=Other">Other users</a>
 	</div>
 
 	<table class="timesheet_table w3-table-all w3-hoverable w3-striped w3-bordered">
 		<thead>
 		<tr class="w3-brown">
-			<th>ID</th>
+			<th>#</th>
 			<th>企業名</th>
 			<th>ユーザー名</th>
-			<th>部署</th>
 			<th>email</th>
-			<th>Dayoff</th>
-			<th>レポート</th>
 			<th></th>
 		</tr>
 		</thead>
@@ -54,16 +61,16 @@
 				<span class="@{{ model.ICON_CLASS }}"></span>
 				&nbsp;
 
-				<a href="{{ $data['url_pattern'] }}/edit/@{{ model.id }}"><span class="glyphicon glyphicon-pencil"></span> <span ng-bind="model.name"></span></a>
+				<a href="{{ $data['url_pattern'] }}/edit/@{{ model.id }}"><span class="fa fa-pencil"></span> <span ng-bind="model.name"></span></a>
 
 			</td>
-			<td><span ng-bind="model.session_name"></span></td>
-			<td><a href="mailto:@{{ model.email }}"><span class="glyphicon glyphicon-envelope"></span> <span ng-bind="model.email"></span></a></td>
-			<td><span ng-bind="model.dayoff"></span></td>
-			<td><a href="/report/project?user_id=@{{ model.id }}"><span class="fa fa-file-o" aria-hidden="true"></span> プロジェクトのレポート</a></td>
+			<td><a href="mailto:@{{ model.email }}"><span class="fa fa-envelope"></span> <span ng-bind="model.email"></span></a></td>
 			<td>
-				<a href="{{ $data['url_pattern'] }}/edit/@{{ model.id }}"><span class="glyphicon glyphicon-pencil"></span></a>
-				| <a href="javascript:void(0);" ng-click="delete_recover(model.id, model.DELETE_FLAG_ACTION)"><i class="@{{model.DELETED_RECOVER_ICON}} @{{model.DELETED_RECOVER_COLOR}}"></i></a>
+				@if(isset($data['permission_flag']) && (in_array($data['permission_flag'], array('Master', 'Owner'))))
+					<a href="{{ $data['url_pattern'] }}/edit/@{{ model.id }}"><span class="fa fa-pencil"></span></a>
+
+					| <a href="javascript:void(0);" ng-click="delete_recover(model.id, model.DELETE_FLAG_ACTION)"><i class="@{{model.DELETED_RECOVER_ICON}} @{{model.DELETED_RECOVER_COLOR}}"></i></a>
+				@endif
 			</td>
 		</tr>
 	</table>
@@ -76,6 +83,6 @@
 
 </div> {{-- <div ng-app="myApp" ng-controller="myCtrl"> --}}
 
-@include('_include.admin_footer', [
+@include('_include.master.footer', [
 	'js_list'				=> true,
 ])
